Change save-creation form to use `react-hook-form` like the settings page
This makes the two forms more similar and improve the quality of the
error message for required fields.
