+++ src/ui/settings/elements.js
@@ -1,81 +1,8 @@
 //; -*- mode: rjsx;-*-
 /** @jsx jsx */
 import { jsx, Styled } from 'theme-ui';
-
-import { useTranslation } from 'react-i18next';
 import { Box } from '@theme-ui/components';
 
-// A styled `<input>`` element.
-export const Input = ({
-  errors,
-  label,
-  name,
-  register,
-  required,
-  validate = undefined,
-  type = 'text',
-  ...rest
-}) => {
-  const { t } = useTranslation();
-
-  return (
-    <div
-      sx={{
-        '&:not(:last-child)': {
-          mb: 3
-        }
-      }}
-    >
-      <label
-        sx={{
-          color: 'text',
-          display: 'block',
-          fontSize: 2,
-          fontWeight: 'bold'
-        }}
-      >
-        {label}
-        <div
-          sx={{
-            display: 'block',
-            boxSizing: 'border-box',
-            clear: 'both',
-            fontSize: 2,
-            position: 'relative',
-            textAlign: 'left'
-          }}
-        >
-          <input
-            aria-invalid={errors[name] ? 'true' : 'false'}
-            aria-describedby={`${name}Error`}
-            autoComplete="off"
-            name={name}
-            ref={register({
-              validate,
-              ...required ? { required: t('forms-validation-error-required') } : {},
-            })}
-            sx={{ variant: 'styles.input' }}
-            type={type}
-            {...rest}
-          />
-          {errors[name] && (
-            <p
-              id={`${name}Error`}
-              sx={{
-                color: 'error',
-                fontSize: 1,
-                fontWeight: 'body',
-                mt: 1
-              }}
-            >
-              {errors[name].message}
-            </p>
-          )}
-        </div>
-      </label>
-    </div>
-  );
-};
 
 export const SettingsSection = ({ title, children }) => (
   <Box
